LayoutTests/fast/text-extraction/debug-text-extraction-shorten-urls.html

<!-- webkit-test-runner [ useFlexibleViewport=true textExtractionEnabled=true dumpJSConsoleLogInStdErr=true ] -->
<!DOCTYPE html>
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta charset="utf-8">
<head>
<style>
body {
    white-space: pre-wrap;
}

.text-representation {
    white-space: pre-wrap;
}

.test-section {
    margin: 20px 0;
    padding: 10px;
    border: 1px solid #ccc;
}

img {
    width: 50px;
    height: 50px;
}

a, img {
    display: block;
}
</style>
<script src="../../resources/ui-helper.js"></script>
</head>
<body>
<main>
    <section class="test-section" title="Links">
        <a href="https://example.com/search?q=webkit&lang=en&filter=recent&sort=relevance&page=1">Multiple query parameters</a>
        <a href="https://example.com/search?q=html&lang=en&client=safari">Multiple query parameter 2</a>
        <a href="https://example.com/path/to/file.html">Path components 1</a>
        <a href="https://example.com/192a2083-6ccc-4508-9b59-6d6f59846dc5/oiiai_cat/19284789123">Path components 2</a>
        <a href="javascript:alert('Hello World')">JavaScript link</a>
        <a href="data:text/html;charset=utf-8,%3Ch1%3EHello%3C%2Fh1%3E">HTML data URL</a>
        <a href="mailto:test@example.com">Email with params</a>
        <a href="mailto:test@example.com">mailto:test@example.com</a>
        <a href="tel:+1-555-123-4567">Phone number</a>
        <a href="blob:https://example.com/550e8400-e29b-41d4-a716-446655440000">Blob URL</a>
        <a href="https://example.com/docs/guide.html?version=latest#installation-steps">Path, query, and fragment</a>
        <a href="https://example.com/docs/guide.html?version=2025">Query parameter</a>
        <a href="https://example.com/docs/guide.html?version=2025">Query parameter (duplicate)</a>
        <a href="https://example.com/d2fb0de6-daf3-42d7-b219-1a243d872156?cid=b3029797">Plain link</a>
        <a href="https://example.com/7e61b3f7-8d81-4225-8d68-f646f977ce33?cid=63e3d9fb0bf5">Plain link (2)</a>
        <a href="https://example.com/b3029797-1fb8-4edf-97bd-63e3d9fb0bf5?cid=4edf97bd">Plain link (3)</a>
        <a href="file:///Users/test/089bac2e873e45608912886c4e028584/cat.png">Local file</a>
        <a href="debug-text-extraction-shorten-urls.html">Self link</a>
    </section>
    <section class="test-section" title="Images">
        <img src="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='100' height='100'%3E%3Ccircle cx='50' cy='50' r='40' fill='red'/%3E%3C/svg%3E" alt="SVG data URL">
        <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg==" alt="PNG data URL">
        <img src="https://cdn.example.com/media/users/12039859134234/uploads/2024/photos/vacation-2024-summer-beach.jpg" alt="Long image path">
        <img src="https://images.example.com/foo/thumbnail.png" alt="Image with path">
        <img src="https://images.example.com/thumbnail.png?id=98765&size=large&quality=high&format=webp&cache=false" alt="Image with query params">
        <img src="https://assets.example.com/61d7035fa8d4.gif" alt="Image with high entropy name">
        <img src="https://assets.example.com/70a27eafd353.gif" alt="Image with high entropy name 2">
        <img src="https://assets.example.com/70a27eafd353.gif" alt="Image with high entropy name (duplicate)">
        <img src="https://assets.example.com/21e0397d-ed0b77b86272" alt="Image with high entropy name, no extension">
        <img src="https://assets.example.com/6539472cv17d.jpg;canvasHeight=130;canvasWidth=130" alt="File extension with trailing text">
    </section>
</main>

<script>
addEventListener("load", async () => {
    if (!window.testRunner)
        return;

    testRunner.dumpAsText();
    testRunner.waitUntilDone();

    const results = [];
    for (let outputFormat of ["texttree", "markdown"]) {
        const heading = document.createElement("h1");
        heading.textContent = `-- ${outputFormat}`;

        const container = document.createElement("pre");
        container.classList.add("text-representation");
        let textContent = await UIHelper.requestDebugText({
            includeURLs: true,
            shortenURLs: true,
            outputFormat,
        });

        container.textContent = textContent;
        results.push(heading);
        results.push(container);
        results.push(document.createElement("br"));
    }

    document.body.replaceChildren(...results);

    testRunner.notifyDone();
});
</script>
</body>
</html>
